Added a scaler and a problem-formatter
The scaler is a method of Problem and scales the component of each
vector to given values. Added also al helper to read the ranges and
values to scale from a libSVM file.

Also added a Problem method to put the Problem as a slice of string
where each string is a vector in the libSVM format:
label index1:value1 index2:value2 etc.